pyproject.toml
[tool.poetry]
name = "automation-read-invoices"
version = "0.1.0"
description = ""
authors = ["Camila Caldas  <camisccb@gmail.com>"]
readme = "README.md"
packages = [{include = "automation_read_invoices"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-telegram-bot = "^20.0"
httpx = "^0.23.3"
pathlib = "^1.0.1"
python-dotenv = "^0.21.0"
google-cloud-vision = "^3.3.1"
autopep8 = "^2.0.1"
google-cloud-storage = "^2.7.0"
google-cloud-bigquery = "^3.6.0"
pandas = "^1.5.3"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
